requirements.txt
Flask==3.0.3
python-dotenv==0.20.0
requests==2.22.0
virtualenv==20.4.2
vonage==3.16.0
pprintpp==0.4.0
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
